The Product Description Component
===
Important command for this module:
ng generate component product-description
===

To verify your work locally, run the following command in a terminal: npm run test:part2. The output will let you know which tasks are passing and failing.

To verify your work on Pluralsight, commit and push your changes to GitHub and click the Check My Work button.

1. Displaying The ProductPageComponent

In app.component.html, replace <h1>{{title}}</h1> with <app-product-page></app-product-page>, which is the selector for the ProductPageComponent.

Remember, you can run ng serve to start the application, and when you visit http://localhost:4200 in a browser, you should see the component’s HTML and CSS rendered on the page.

2. Creating The ProductDescriptionComponent

Let’s create a new component by typing ng generate component product-description in the command line.

3. Moving The Product Description HTML

Look in product-page.component.html for the div with a class of description and move that div tag and all of its children to the product-description.component.html file, replacing any HTML that’s already in the ProductDescriptionComponent file.

4. Inserting The product-description Tag

In product-page.component.html in the empty space where the old div was, add an opening and closing tag for the app-product-description selector. Once you complete this task, the page should render the same in the browser, but now the HTML is being loaded from the ProductDescriptionComponent.

5. Setting Font Styles In The Description

In product-description.component.css, select the p tag and set the font-size to 16px, the font-family to Helvetica with a fallback of Arial and sans-serif, and a font-weight of normal.

6. Setting Image Styles In The Description

Still in product-description.component.css, select the img tag and set the width to 100%.